Python_basic_10_08_2023/requirements.txt
flask
transformers
torch
pyngrok
flask
flask-ngrok
Flask-Cors
python-dotenv
flask-socketio
simple-websocket
python-socketio